requirements.txt
streamlit>=1.31.0
openai>=1.12.0
qdrant-client>=1.7.0
python-dotenv>=1.0.0
tqdm>=4.66.0
pandas==2.2.0
numpy>=2.2.0
tiktoken>=0.6.0 